NPC movement
Changed up how the NPC moves:
Now there's a prefab called 'NPC Destination' that can be placed on the
field at will. The object is then added to the NPC's path array, which
it follows when the boolean npcMovement is true. Paths are looped by
checking loopPath on the NPC.