New Spell Animations "Heal and Wind"
I have added Temp animations for the spells Heal and Wind. Which are
self imposed spells. One obviously healing you, and the other one acting
as a Haste like spell that increases your speed, but doesn't harm you if
you run in to walls with increased speed.

+Temp Wind Spell Animation
+Temp Heal Spell Animation
+New Sprite Sheet - Temp Heal and Wind Animation
! Heal and Wind Spells are TEMP! And are not Final.

In other news, Sarah's New sprite is about 30% done. God it's taking a
while...