Money HUD
There's now a Wallet (Money) in the top right corner. And of course
we're using G, because we can't be creative or anything.

+ HUD Money.

~ (Probably) fixed a bug where the enemy would try to backtrace when it
didn't have a point to go back to. This problem should be done. Because
of how programming works, this probably introduced like six or seven NEW
bugs, so try to find those please.